MALI-109 Fixes
- separate miniappvc subclasses
- localize strings
- run pod deintegrate
