+++ MiniApp/Classes/ui/MiniAppViewController+ViewState.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+// MARK: - ViewState
+extension MiniAppViewController {
+
+    enum ViewState {
+        case loading
+        case error
+        case success
+    }
+
+}
