+++ MiniApp/en.lproj/Localizable.strings
@@ -37,3 +37,8 @@
 "wait_message" = "Please wait...";
 "ok" = "OK";
 "cancel" = "Cancel";
+
+// MARK: - UI
+"ui_close_button_title" = "Close";
+"ui_loading_failed_error" = "An error has occured.\nPlease try again";
+"ui_retry_button_title" = "Retry";
